imdb geliyor

+++ movie/views.py
@@ -4,17 +4,29 @@ from django.shortcuts import render_to_response, render
 from models import *
 from forms import UnratedForm, RatingForm, RecommendationForm
 from django.contrib.auth.decorators import login_required
-from django.template import RequestContext
 from django.forms.formsets import formset_factory
-from django.core.context_processors import csrf
 from django.forms.models import modelformset_factory
-import json, sys
+import json
 from webservice import WebService
+import random
 
 #constants
 context = os.path.basename(os.path.dirname(os.path.realpath(__file__)))
 
 
+@login_required(login_url='/login/')
+def user_popover(request):
+    user_id = request.GET.get('user')
+    user = User.objects.get(pk=user_id)
+    sim_perc = random.randint(0,100)
+    try:
+        f = Follow.objects.get(follower=request.user,
+                followee=user)
+    except Follow.DoesNotExist:
+        f=Follow(follower=request.user,followee=user)
+
+    return render(request, context + '/user_popover.html',
+            {'context':context,'user':user,'sim_perc':sim_perc,'f':f})
 @login_required(login_url='/login/')
 def search(request):
     items = Item.objects.filter( name__icontains=request.GET.get('q'))[0:20]
@@ -42,10 +54,94 @@ def reclist(request):
     return HttpResponse(json.dumps(reclist),mimetype="application/json")
 
 @login_required(login_url='/login/')
-def home(request):
+def home_comments(request):
+    item_id = request.GET.get('itemid')
+    item = Item.objects.get(pk=int(item_id))
+    if request.GET.get('comment',False):
+        comment = request.GET.get('comment')
+        Comment.objects.create(user=request.user,item=item,comment=comment)
+
+    comments = Comment.objects.filter(item=item)
+    return render(request, context + '/home_comments.html',{'context':context,
+        'comments':comments,'item':item})
+@login_required(login_url='/login/')
+def _home_more(request):
+    offset = int(request.GET.get('offset'))
+    limit = 10
     followees = [f.followee_id for f in Follow.objects.filter(follower=request.user)]
-    actions = Action.objects.filter(user__in=followees).order_by('-when')
-    return render(request, context + '/home.html',{'context':context,'actions':actions,'fusers':getneighbors(request.user)})
+    actions = Action.objects.filter(user__in=followees).order_by('-when')[offset:offset+limit]
+    tuples = []
+    for ac in actions:
+        r = {}
+        r['act'] = {'what': ac.what, 'when': str(ac.when), 'username': ac.user.username, 'pk': ac.user.pk, 'pic_url': ac.user.profile.pic_url}
+        if ac.what == 'follow':
+            user = User.objects.get(pk=ac.gen_id) 
+            r['user'] = {'pk': user.pk, 'username': user.username, 'pic_url': str(user.profile.pic_url)}
+            items = Item.objects.filter()[:5]
+            r['items'] = [(item.pk, item.name, item.year) for item in items]
+        elif ac.what == 'rate':
+            item = Item.objects.get(pk=ac.gen_id)
+            r['item'] = {'pk': item.pk, 'year': item.year, 'name': item.name, 'rating':Rating.objects.get(user=ac.user,item=item).rating, 'pre':3}
+        elif ac.what == 'recommend':
+            item = Item.objects.get(pk=ac.gen_id)
+            r['item'] = {'pk': item.pk, 'name': item.name, 'year': item.year}
+
+        tuples.append(r)
+
+
+
+    return HttpResponse(json.dumps(tuples), mimetype="application/json")
+    #return HttpResponse(json.dumps(tuples) )
+    a = [{'act': {'user': ac.user.username}, 'obj':2}, {'act': 7, 'obj': 4}]
+    a = [{'act': tuples, 'obj':2}, {'act': 7, 'obj': 4}]
+
+    return HttpResponse(json.dumps(a), mimetype="application/json")
+    #return HttpResponse("Offset is " + str(tuples))
+
+@login_required(login_url='/login/')
+def home(request):
+    offset = int(request.GET.get('offset',0))
+    limit = 10
+    followees = [f.followee_id for f in Follow.objects.filter(follower=request.user)]
+    actions = Action.objects.filter(user__in=followees).order_by('-when')[offset:limit]
+    tuples = []
+    for ac in actions:
+        if ac.what == 'follow':
+            tuples.append((ac,
+                {'user':User.objects.get(pk=ac.gen_id),'items':Item.objects.filter()[:5]}
+                ))
+        elif ac.what == 'rate':
+            item = Item.objects.get(pk=ac.gen_id)
+            tuples.append((ac,
+                {'item': item,'rating':Rating.objects.get(user=ac.user,item=item).rating,'pre':3}
+                ))
+        elif ac.what == 'recommend':
+            tuples.append((ac,Item.objects.get(pk=ac.gen_id)))
+
+    return render(request, context + '/home_orig.html',{'context':context,'tuples':tuples, 'noffset':limit+offset,'fusers':getneighbors(request.user)})
+
+@login_required(login_url='/login/')
+def home_more(request):
+    offset = int(request.GET.get('offset',0))
+    limit = 10
+    followees = [f.followee_id for f in Follow.objects.filter(follower=request.user)]
+    actions = Action.objects.filter(user__in=followees).order_by('-when')[offset:offset+limit]
+    tuples = []
+    for ac in actions:
+        if ac.what == 'follow':
+            tuples.append((ac,
+                {'user':User.objects.get(pk=ac.gen_id),'items':Item.objects.filter()[:5]}
+                ))
+        elif ac.what == 'rate':
+            item = Item.objects.get(pk=ac.gen_id)
+            tuples.append((ac,
+                {'item': item,'rating':Rating.objects.get(user=ac.user,item=item).rating,'pre':3}
+                ))
+        elif ac.what == 'recommend':
+            tuples.append((ac,Item.objects.get(pk=ac.gen_id)))
+
+    return render(request, context + '/home_more.html',{'context':context,'tuples':tuples, 'noffset':limit+offset})
+
 
 @login_required(login_url='/login/')
 def rate(request):
@@ -53,8 +149,11 @@ def rate(request):
     i = Item.objects.get(pk=request.GET.get('item_id'))
     r = int(request.GET.get('rating'))
     rating = Rating.objects.filter(user=u,item=i)
-    rating.delete()
-    rating=None
+    if r < 0:
+        Action.objects.filter(user=u,what='rate',gen_id=i.pk).delete()
+        rating.delete()
+        rating=None
+
     if r>0:
         rating = Rating.objects.create(user=u,item=i,rating=r)
         Action.objects.create(user=request.user,what='rate',
@@ -136,14 +235,68 @@ def detail(request,pk):
     from django.db.models import Avg, Count
     item = Item.objects.get(pk=pk)
     res = Rating.objects.filter(item_id=item.pk).aggregate(num_ratings=Count('id'),avg_rating=Avg('rating'))
-    if UserRec.objects.filter(user=request.user,
+    '''if UserRec.objects.filter(user=request.user,
             item=item).count():
         userrec=UserRec.objects.get(user=request.user,item=item)
     else:
-        userrec=None
+        userrec=None'''
+
+    if Rating.objects.filter(user=request.user,item=item).exists():
+        rating = Rating.objects.get(user=request.user,item=item)
+    else:
+        rating = Rating(user=request.user,item=item,rating=random.randint(1,5))
+    row = (item,rating)
+
+    ratingc = Rating.objects.filter(item=item).count()
+    reviewc = Comment.objects.filter(item=item).count()
+
+    comments = Comment.objects.filter(item=item)
+    
+    sim_items = random.sample(Item.objects.all(), 6)
+    
 
     return render(request,context + '/item.html',{
-        'status':str(res),'item':item,'userrec':userrec})
+        'context':context,'status':str(res),'item':item,
+        'row':row,'ratingc':ratingc,'reviewc':reviewc,'rlist':xrange(1,6),
+        'comments':comments, 'sim_items':sim_items})
+    #return render(request,context + '/item.html',{})
+
+@login_required(login_url='/login/')
+def profile_ratings(request):
+    user = request.user
+    ratings = Rating.objects.filter(user=user)
+    rows =[]
+    for r in ratings:
+        rows.append((r.item,r))
+
+    return render(request, context + '/profile_ratings.html', {
+        'context':context,'rows':rows,
+        })
+@login_required(login_url='/login/')
+def profile_users(request):
+    from random import randint
+    if request.GET.get('u'):
+        user = User.objects.get(pk=request.GET.get('u'))
+    else:
+        user = request.user
+    if request.GET.get('t',False) == 'follows':
+        followings = Follow.objects.filter(follower=user)
+        users = [f.followee for f in followings]
+    elif request.GET.get('t',False) == 'followers':
+        followings = Follow.objects.filter(followee=user)
+        users = [f.follower for f in followings]
+    rows=[]
+    for u in users:
+        try:
+            f = Follow.objects.get(follower=request.user,followee=u)
+        except Follow.DoesNotExist:
+            f = Follow(follower=request.user,followee=u)
+        rows.append((u,f,randint(0,100)))
+
+    return render(request, context + '/profile_users.html', {
+        'context':context,'rows': rows, 
+        'auser': user,'user':request.user,
+        })
 
 @login_required(login_url='/login/')
 def profile(request):
@@ -158,25 +311,47 @@ def profile(request):
     for r in ratings:
         rows.append((r.item,r))
     
-    followings = Follow.objects.filter(follower=user)
-    followees = [f.followee for f in followings]
+    followees = Follow.objects.filter(follower=user)
+    followers = Follow.objects.filter(followee=user)
     try:
         f = Follow.objects.get(follower=request.user,
                 followee=user)
     except Follow.DoesNotExist:
         f=Follow(follower=request.user,followee=user)
 
+    sim_users = getneighbors(user, 12)
+
     recs = UserRec.objects.filter(user=user)
+    offset = int(request.GET.get('offset',0))
+    limit = 10
+    actions = Action.objects.filter(user=user).order_by('-when')[offset:offset+limit]
+    tuples = []
+    for ac in actions:
+        if ac.what == 'follow':
+            tuples.append((ac,
+                {'user':User.objects.get(pk=ac.gen_id),'items':Item.objects.filter()[:5]}
+                ))
+        elif ac.what == 'rate':
+            item = Item.objects.get(pk=ac.gen_id)
+            tuples.append((ac,
+                {'item': item,'rating':Rating.objects.get(user=ac.user,item=item).rating,'pre':3}
+                ))
+        elif ac.what == 'recommend':
+            tuples.append((ac,Item.objects.get(pk=ac.gen_id)))
+
+
     return render(request, context + '/profile.html', {
-        'context':context,'recs': recs, 'followees': followees, 
+        'context':context,'recs': recs, 'followees': followees,'followers':followers, 
         'auser': user,'user':request.user, 'f': f,'rows':rows,'rlist':range(1,6),
+        'sim_users': sim_users, 'tuples' : tuples,'noffset':limit+offset,
         })
 
 
 #### UTILS ######
 
 
-def getneighbors(user):
+def getneighbors(user, n=5):
+    return User.objects.filter()[:n]
     w = WebService(context)
     strarr = w.get_nearestneighbors(user.id)
     uids = [int(s.split(';')[0]) for s in strarr]
